put everything in examples
