+++ rlberry/agents/bandits/__init__.py
@@ -2,4 +2,3 @@ from .bandit_base import BanditWithSimplePolicy
 from .index_agents import IndexAgent, RecursiveIndexAgent
 from .randomized_agents import RandomizedAgent
 from .thompson_sampling import TSAgent
-from .bai_agents import SeqHalvAgent
