+++ rlberry/envs/bandits/__init__.py
@@ -1,4 +1,3 @@
 from .bandit_base import AdversarialBandit, Bandit
 from .corrupted_bandits import CorruptedLaws, CorruptedNormalBandit
 from .stochastic_bandits import BernoulliBandit, NormalBandit
-from .mirror_env import MirrorBandit
